Test of sleeping functionality in MQTT-SN
